flight_tracking/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
Django==4.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
psycopg2>=2.8
gunicorn==20.1.0
flake8==4.0.1
black==21.12b0
requests==2.27.1
django-cors-headers==3.12.0
pandas==1.4.2
dask==2022.5.2
boto3==1.24.36
airportsdata==20220921